Wrap ruamel yaml library import due to type support
